+++ lib/cli/run-helpers.js
@@ -9,11 +9,9 @@
 
 const fs = require('fs');
 const path = require('path');
-const ansi = require('ansi-colors');
 const debug = require('debug')('mocha:cli:run:helpers');
-const minimatch = require('minimatch');
-const utils = require('../utils');
 const watchRun = require('./watch-run');
+const collectFiles = require('./collect-files');
 
 const cwd = (exports.cwd = process.cwd());
 
@@ -94,90 +92,18 @@ exports.handleRequires = (requires = []) => {
 };
 
 /**
- * Smash together an array of test files in the correct order
- * @param {Object} [opts] - Options
- * @param {string[]} [opts.extension] - File extensions to use
- * @param {string[]} [opts.spec] - Files, dirs, globs to run
- * @param {string[]} [opts.ignore] - Files, dirs, globs to ignore
- * @param {boolean} [opts.recursive=false] - Find files recursively
- * @param {boolean} [opts.sort=false] - Sort test files
- * @returns {string[]} List of files to test
- * @private
- */
-exports.handleFiles = ({
-  ignore = [],
-  extension = [],
-  file = [],
-  recursive = false,
-  sort = false,
-  spec = []
-} = {}) => {
-  let files = [];
-  const unmatched = [];
-  spec.forEach(arg => {
-    let newFiles;
-    try {
-      newFiles = utils.lookupFiles(arg, extension, recursive);
-    } catch (err) {
-      if (err.code === 'ERR_MOCHA_NO_FILES_MATCH_PATTERN') {
-        unmatched.push({message: err.message, pattern: err.pattern});
-        return;
-      }
-
-      throw err;
-    }
-
-    if (typeof newFiles !== 'undefined') {
-      if (typeof newFiles === 'string') {
-        newFiles = [newFiles];
-      }
-      newFiles = newFiles.filter(fileName =>
-        ignore.every(pattern => !minimatch(fileName, pattern))
-      );
-    }
-
-    files = files.concat(newFiles);
-  });
-
-  if (!files.length) {
-    // give full message details when only 1 file is missing
-    const noneFoundMsg =
-      unmatched.length === 1
-        ? `Error: No test files found: ${JSON.stringify(unmatched[0].pattern)}` // stringify to print escaped characters raw
-        : 'Error: No test files found';
-    console.error(ansi.red(noneFoundMsg));
-    process.exit(1);
-  } else {
-    // print messages as an warning
-    unmatched.forEach(warning => {
-      console.warn(ansi.yellow(`Warning: ${warning.message}`));
-    });
-  }
-
-  const fileArgs = file.map(filepath => path.resolve(filepath));
-  files = files.map(filepath => path.resolve(filepath));
-
-  // ensure we don't sort the stuff from fileArgs; order is important!
-  if (sort) {
-    files.sort();
-  }
-
-  // add files given through --file to be ran first
-  files = fileArgs.concat(files);
-  debug('files (in order): ', files);
-  return files;
-};
-
-/**
- * Give Mocha files and tell it to run
+ * Collect test files and run mocha instance.
  * @param {Mocha} mocha - Mocha instance
- * @param {Options} [opts] - Options
- * @param {string[]} [opts.files] - List of test files
+ * @param {Options} [opts] - Command line options
  * @param {boolean} [opts.exit] - Whether or not to force-exit after tests are complete
+ * @param {Object} fileCollectParams - Parameters that control test
+ *   file collection. See `lib/cli/collect-files.js`.
  * @returns {Runner}
  * @private
  */
-exports.singleRun = (mocha, {files = [], exit = false} = {}) => {
+exports.singleRun = (mocha, {exit}, fileCollectParams) => {
+  const files = collectFiles(fileCollectParams);
+  debug('running tests with files', files);
   mocha.files = files;
   return mocha.run(exit ? exitMocha : exitMochaLater);
 };
@@ -185,24 +111,36 @@ exports.singleRun = (mocha, {files = [], exit = false} = {}) => {
 /**
  * Actually run tests
  * @param {Mocha} mocha - Mocha instance
- * @param {Object} [opts] - Options
- * @param {boolean} [opts.watch=false] - Enable watch mode
- * @param {string[]} [opts.extension] - List of extensions to watch
- * @param {string|RegExp} [opts.grep] - Grep for test titles
- * @param {string} [opts.ui=bdd] - User interface
- * @param {boolean} [opts.exit=false] - Force-exit Mocha when tests done
- * @param {string[]} [files] - Array of test files
+ * @param {Object} opts - Command line options
  * @private
  */
-exports.runMocha = (
-  mocha,
-  {watch = false, extension = [], grep = '', ui = 'bdd', exit = false} = {},
-  files = []
-) => {
+exports.runMocha = (mocha, options) => {
+  const {
+    watch = false,
+    extension = [],
+    grep = '',
+    ui = 'bdd',
+    exit = false,
+    ignore = [],
+    file = [],
+    recursive = false,
+    sort = false,
+    spec = []
+  } = options;
+
+  const fileCollectParams = {
+    ignore,
+    extension,
+    file,
+    recursive,
+    sort,
+    spec
+  };
+
   if (watch) {
-    watchRun(mocha, {extension, grep, ui, files});
+    watchRun(mocha, {ui, grep}, fileCollectParams);
   } else {
-    exports.singleRun(mocha, {files, exit});
+    exports.singleRun(mocha, {exit}, fileCollectParams);
   }
 };
