+++ lib/cli/run.js
@@ -16,7 +16,6 @@ const {
 
 const {
   list,
-  handleFiles,
   handleRequires,
   validatePlugin,
   runMocha
@@ -290,8 +289,5 @@ exports.builder = yargs =>
 exports.handler = argv => {
   debug('post-yargs config', argv);
   const mocha = new Mocha(argv);
-  const files = handleFiles(argv);
-
-  debug('running tests with files', files);
-  runMocha(mocha, argv, files);
+  runMocha(mocha, argv);
 };
