+++ lib/cli/watch-run.js
@@ -3,6 +3,7 @@
 const utils = require('../utils');
 const Context = require('../context');
 const Mocha = require('../mocha');
+const collectFiles = require('./collect-files');
 
 /**
  * Exports the `watchRun` function that runs mocha in "watch" mode.
@@ -15,14 +16,16 @@ const Mocha = require('../mocha');
  * Run Mocha in "watch" mode
  * @param {Mocha} mocha - Mocha instance
  * @param {Object} opts - Options
- * @param {string[]} opts.extension - List of extensions to watch
  * @param {string|RegExp} opts.grep - Grep for test titles
  * @param {string} opts.ui - User interface
- * @param {string[]} opts.files - Array of test files
+ * @param {Object} fileCollectParams - Parameters that control test
+ *   file collection. See `lib/cli/collect-files.js`.
+ * @param {string[]} fileCollectParams.extension - List of extensions to watch
  * @private
  */
-module.exports = (mocha, {extension, grep, ui, files}) => {
+module.exports = (mocha, {grep, ui}, fileCollectParams) => {
   let runner;
+  const files = collectFiles(fileCollectParams);
 
   console.log();
   hideCursor();
@@ -34,7 +37,7 @@ module.exports = (mocha, {extension, grep, ui, files}) => {
     process.exit(128 + 2);
   });
 
-  const watchFiles = utils.files(process.cwd(), extension);
+  const watchFiles = utils.files(process.cwd(), fileCollectParams.extension);
   let runAgain = false;
 
   const loadAndRun = () => {
